+++ magik-language-server/src/test/java/nl/ramsolutions/sw/magik/languageserver/hover/HoverProviderTest.java
@@ -77,7 +77,6 @@ class HoverProviderTest {
                 ExemplarDefinition.Sort.SLOTTED,
                 hoverMeTypeRef,
                 Collections.emptyList(),
-                Collections.emptyList(),
                 Collections.emptyList()));
 
         final String code = ""
@@ -142,7 +141,7 @@ class HoverProviderTest {
         final Hover hover = this.provideHover(code, position, definitionKeeper);
         final MarkupContent content = hover.getContents().getRight();
         assertThat(content.getKind()).isEqualTo(MarkupKind.MARKDOWN);
-        assertThat(content.getValue()).contains("Unknown method hover_me() on type sw:integer");
+        assertThat(content.getValue()).isEqualTo("");  // No data.
     }
 
     @Test