+++ magik-squid/src/main/java/nl/ramsolutions/sw/magik/parser/TypeStringBuilderVisitor.java
@@ -3,7 +3,6 @@ package nl.ramsolutions.sw.magik.parser;
 import com.sonar.sslr.api.AstNode;
 import com.sonar.sslr.api.AstNodeType;
 import com.sonar.sslr.api.AstVisitor;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -21,7 +20,7 @@ public final class TypeStringBuilderVisitor implements AstVisitor {
 
     private final Map<AstNode, TypeString> mapping = new HashMap<>();
     private final String currentPakkage;
-    private AstNode topAst;
+    private AstNode topNode;
 
     /**
      * Constructor.
@@ -33,7 +32,7 @@ public final class TypeStringBuilderVisitor implements AstVisitor {
 
     @CheckForNull
     public TypeString getTypeString() {
-        return this.mapping.get(this.topAst);
+        return this.mapping.get(this.topNode);
     }
 
     @Override
@@ -43,116 +42,127 @@ public final class TypeStringBuilderVisitor implements AstVisitor {
             TypeStringGrammar.TYPE_CLONE,
             TypeStringGrammar.TYPE_SELF,
             TypeStringGrammar.TYPE_PARAMETER_REFERENCE,
-            TypeStringGrammar.TYPE_GENERIC,
+            TypeStringGrammar.TYPE_GENERIC_DEFINITION,
+            TypeStringGrammar.TYPE_GENERIC_REFERENCE,
             TypeStringGrammar.TYPE_IDENTIFIER,
             TypeStringGrammar.TYPE_STRING,
             TypeStringGrammar.SYNTAX_ERROR);
     }
 
     @Override
-    public void visitFile(final @Nullable AstNode ast) {
-        this.topAst = ast;
+    public void visitFile(final @Nullable AstNode node) {
+        this.topNode = node;
     }
 
     @Override
-    public void leaveFile(final @Nullable AstNode ast) {
+    public void leaveFile(final @Nullable AstNode node) {
         // Pass.
     }
 
     @Override
-    public void visitNode(final AstNode ast) {
+    public void visitNode(final AstNode node) {
         // Pass.
     }
 
     @Override
-    public void leaveNode(final AstNode ast) {
-        if (ast.is(TypeStringGrammar.TYPE_UNDEFINED)) {
-            this.buildUndefined(ast);
-        } else if (ast.is(TypeStringGrammar.TYPE_CLONE, TypeStringGrammar.TYPE_SELF)) {
-            this.buildSelf(ast);
-        } else if (ast.is(TypeStringGrammar.TYPE_PARAMETER_REFERENCE)) {
-            this.buildParameterRef(ast);
-        } else if (ast.is(TypeStringGrammar.TYPE_GENERIC)) {
-            this.buildGeneric(ast);
-        } else if (ast.is(TypeStringGrammar.TYPE_IDENTIFIER)) {
-            this.buildIdentifier(ast);
-        } else if (ast.is(TypeStringGrammar.TYPE_STRING)) {
-            this.buildTypeString(ast);
-        } else if (ast.is(TypeStringGrammar.SYNTAX_ERROR)) {
-            this.buildUndefined(ast);
+    public void leaveNode(final AstNode node) {
+        if (node.is(TypeStringGrammar.TYPE_UNDEFINED)) {
+            this.buildUndefined(node);
+        } else if (node.is(TypeStringGrammar.TYPE_CLONE, TypeStringGrammar.TYPE_SELF)) {
+            this.buildSelf(node);
+        } else if (node.is(TypeStringGrammar.TYPE_PARAMETER_REFERENCE)) {
+            this.buildParameterRef(node);
+        } else if (node.is(TypeStringGrammar.TYPE_GENERIC_DEFINITION)) {
+            this.buildGenericDefinition(node);
+        } else if (node.is(TypeStringGrammar.TYPE_GENERIC_REFERENCE)) {
+            this.buildGenericReference(node);
+        } else if (node.is(TypeStringGrammar.TYPE_IDENTIFIER)) {
+            this.buildIdentifier(node);
+        } else if (node.is(TypeStringGrammar.TYPE_STRING)) {
+            this.buildTypeString(node);
+        } else if (node.is(TypeStringGrammar.SYNTAX_ERROR)) {
+            this.buildUndefined(node);
         } else {
-            throw new IllegalStateException("Unknown node type: " + ast.getType());
+            throw new IllegalStateException("Unknown node type: " + node.getType());
         }
     }
 
-    private void buildUndefined(final AstNode ast) {
+    private void buildUndefined(final AstNode node) {
         final TypeString part = TypeString.UNDEFINED;
 
-        this.mapping.put(ast, part);
+        this.mapping.put(node, part);
     }
 
-    private void buildSelf(final AstNode ast) {
+    private void buildSelf(final AstNode node) {
         final TypeString part = TypeString.SELF;
 
-        this.mapping.put(ast, part);
+        this.mapping.put(node, part);
     }
 
-    private void buildParameterRef(final AstNode ast) {
-        final List<AstNode> childAsts = ast.getChildren();
+    private void buildParameterRef(final AstNode node) {
+        final List<AstNode> childAsts = node.getChildren();
         final AstNode identifierAst = childAsts.get(2);
         final String refStr = identifierAst.getTokenValue();
         final TypeString part = TypeString.ofParameterRef(refStr);
 
-        this.mapping.put(ast, part);
+        this.mapping.put(node, part);
     }
 
-    private void buildGeneric(final AstNode ast) {
-        final List<AstNode> childAsts = ast.getChildren();
-        final AstNode identifierAst = childAsts.get(1);
-        final String str = identifierAst.getTokenValue();
-        final TypeString part = TypeString.ofGeneric(str);
+    private void buildGenericDefinition(final AstNode node) {
+        final AstNode identifierNode = node.getFirstChild(TypeStringGrammar.TYPE_IDENTIFIER);
+        final String identifier = identifierNode.getTokenValue();
+        final AstNode typeStringNode = node.getFirstChild(TypeStringGrammar.TYPE_STRING);
+        final TypeString typeString = this.mapping.get(typeStringNode);
+        final TypeString part = TypeString.ofGenericDefinition(identifier, typeString);
 
-        this.mapping.put(ast, part);
+        this.mapping.put(node, part);
     }
 
-    private void buildIdentifier(final AstNode ast) {
-        final String str = ast.getTokenValue();
-        final AstNode genericDefinitionsNode = ast.getFirstChild(TypeStringGrammar.TYPE_GENERIC_DEFINITIONS);
-        final List<TypeString> generics = genericDefinitionsNode != null
-            ? genericDefinitionsNode.getChildren(TypeStringGrammar.TYPE_STRING).stream()
-                .map(AstNode::getFirstChild)
-                .map(this.mapping::get)
-                .map(Objects::requireNonNull)
-                .collect(Collectors.toList())
-            : Collections.emptyList();
-        final TypeString[] genericsArr = generics.toArray(TypeString[]::new);
+    private void buildGenericReference(final AstNode node) {
+        final AstNode identifierNode = node.getFirstChild(TypeStringGrammar.SIMPLE_IDENTIFIER);
+        final String identifier = identifierNode.getTokenValue();
+        final TypeString part = TypeString.ofGenericReference(identifier);
+
+        this.mapping.put(node, part);
+    }
+
+    private void buildIdentifier(final AstNode node) {
+        final String str = node.getTokenValue();
+        final List<AstNode> genericNodes = node.getChildren(
+            TypeStringGrammar.TYPE_GENERIC_DEFINITION,
+            TypeStringGrammar.TYPE_GENERIC_REFERENCE);
+        final TypeString[] genericsArr = genericNodes.stream()
+            .map(this.mapping::get)
+            .collect(Collectors.toList())
+            .toArray(TypeString[]::new);
         final TypeString part = TypeString.ofIdentifier(str, this.currentPakkage, genericsArr);
 
-        this.mapping.put(ast, part);
+        this.mapping.put(node, part);
     }
 
-    private void buildTypeString(final AstNode ast) {
-        final List<AstNode> childAsts = ast.getChildren(
+    private void buildTypeString(final AstNode node) {
+        final List<AstNode> childNodes = node.getChildren(
             TypeStringGrammar.TYPE_UNDEFINED,
             TypeStringGrammar.TYPE_CLONE,
             TypeStringGrammar.TYPE_SELF,
             TypeStringGrammar.TYPE_PARAMETER_REFERENCE,
-            TypeStringGrammar.TYPE_GENERIC,
+            TypeStringGrammar.TYPE_GENERIC_DEFINITION,
+            TypeStringGrammar.TYPE_GENERIC_REFERENCE,
             TypeStringGrammar.TYPE_IDENTIFIER,
             TypeStringGrammar.SYNTAX_ERROR);
-        final List<TypeString> childTypeStrings = childAsts.stream()
+        final List<TypeString> childTypeStrings = childNodes.stream()
             .map(this.mapping::get)
             .map(Objects::requireNonNull)
             .collect(Collectors.toList());
-        if (childAsts.isEmpty()) {
+        if (childNodes.isEmpty()) {
             throw new IllegalStateException();
         }
 
         final TypeString[] childTypeStringsArr = childTypeStrings.toArray(TypeString[]::new);
-        final TypeString part = childAsts.size() == 1
+        final TypeString part = childNodes.size() == 1
             ? childTypeStrings.get(0)
             : TypeString.ofCombination(this.currentPakkage, childTypeStringsArr);
-        this.mapping.put(ast, part);
+        this.mapping.put(node, part);
     }
 
 }